requirements.txt
tensorflow==2.15.0
keras==2.15.0
numpy==1.26.4
pandas==2.0.3
scikit-learn==1.3.2
matplotlib==3.8.0
seaborn==0.13.0
Pillow==10.0.0
streamlit==1.29.0